[Core/Quests] - All The Core Scripts For the Commited Quest SQL Files

+++ src/server/scripts/Northrend/borean_tundra.cpp
@@ -2515,17 +2515,19 @@ public:
 			if (GetTargetApplication()->GetRemoveMode() != AURA_REMOVE_BY_EXPIRE || !GetCaster())
 				return;
 
-			Creature* owner = GetOwner()->ToCreature();
-			owner->RemoveAllAurasExceptType(SPELL_AURA_DUMMY);
-			owner->CombatStop(true);
-			owner->DeleteThreatList();
-			owner->GetMotionMaster()->Clear(false);
-			owner->GetMotionMaster()->MoveFollow(GetCaster(), 4.0f, 0.0f);
-			owner->CastSpell(owner, SPELL_SUBDUED, true);
-			GetCaster()->CastSpell(GetCaster(), SPELL_DRAKE_HATCHLING_SUBDUED, true);
-			owner->setFaction(35);
-			owner->SetFlag(UNIT_FIELD_FLAGS, UNIT_FLAG_IMMUNE_TO_PC | UNIT_FLAG_IMMUNE_TO_NPC);
-			owner->DespawnOrUnsummon(3 * MINUTE*IN_MILLISECONDS);
+			if (Creature* owner = GetOwner()->ToCreature())
+			{
+				owner->RemoveAllAurasExceptType(SPELL_AURA_DUMMY);
+				owner->CombatStop(true);
+				owner->DeleteThreatList();
+				owner->GetMotionMaster()->Clear(false);
+				owner->GetMotionMaster()->MoveFollow(GetCaster(), 4.0f, 0.0f);
+				owner->CastSpell(owner, SPELL_SUBDUED, true);
+				GetCaster()->CastSpell(GetCaster(), SPELL_DRAKE_HATCHLING_SUBDUED, true);
+				owner->setFaction(35);
+				owner->SetFlag(UNIT_FIELD_FLAGS, UNIT_FLAG_IMMUNE_TO_PC | UNIT_FLAG_IMMUNE_TO_NPC);
+				owner->DespawnOrUnsummon(3 * MINUTE*IN_MILLISECONDS);
+			}
 		}
 
 		void Register()
